Move page doc helpers out of library code, and revert preserveWhitespace
We do not use 'preserveWhitespace: full' here to allow us to format HTML in fixtures without breaking tests - see  in ./cypress/helpers/editor.ts
